move api tokens models to schema pkg
