+++ schema/accessToken.go
@@ -1,4 +1,4 @@
-package models
+package schema
 
 import (
 	"context"
